refactor: wrapping halo plugin manager as a thread safe class
